job-scraper-service/requirements.txt
beautifulsoup4==4.13.3
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
Flask==3.1.0
html5lib==1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
requests==2.32.3
schedule==1.2.2
six==1.17.0
soupsieve==2.6
typing_extensions==4.12.2
urllib3==2.3.0
webencodings==0.5.1
Werkzeug==3.1.3
Flask-SQLAlchemy